JDP230401-33 v2 - Fix ProductDto

+++ src/main/java/com/kodilla/ecommercee/service/ProductDbService.java
@@ -1,7 +1,10 @@
 package com.kodilla.ecommercee.service;
 
+import com.kodilla.ecommercee.controller.exceptions.ProductGroupNotFoundException;
 import com.kodilla.ecommercee.controller.exceptions.ProductNotFoundException;
 import com.kodilla.ecommercee.domain.Product;
+import com.kodilla.ecommercee.domain.ProductGroup;
+import com.kodilla.ecommercee.repository.ProductGroupRepository;
 import com.kodilla.ecommercee.repository.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +16,7 @@ import java.util.List;
 public class ProductDbService {
 
     private final ProductRepository productRepository;
+    private final ProductGroupRepository groupRepository;
 
     public List<Product> getProductList() {
         return productRepository.findAll();
@@ -26,7 +30,14 @@ public class ProductDbService {
         return productRepository.save(product);
     }
 
-    public void removeProduct(final Long productId) {
+    public void deleteProductById(final Long productId) throws ProductNotFoundException {
+        if (!productRepository.existsById(productId)){
+            throw new ProductNotFoundException();
+        }
         productRepository.deleteById(productId);
+    }
+
+    public ProductGroup getGroup(Long groupId) throws ProductGroupNotFoundException {
+        return groupRepository.findById(groupId).orElseThrow(ProductGroupNotFoundException::new);
     }
 }
